requirements.txt
streamlit==1.32.0
openai==1.12.0
pymongo==4.6.1
requests==2.31.0
python-dotenv==1.0.0
langchain==0.3.27
langchain-openai==0.3.33